Keep track of which items were imported
This allows us to safely remove all of those on a refresh, which
avoids dupes if the X menu entry changes slightly (for example, if
a parameter to the command gets quoted).

Maintains compatbility with old menu files, although it can't fix
any existing dupes because there's no way to retroactively determine
if an old entry was imported.  It does mark items as imported on a
refresh however, so it should help with any future dupes.