Add migrate-settings script
Unfortunately, it doesn't seem to be possible to load some old
settings files using PyQt5. This is because that version removed the
QString class which we were using to store our icon theme setting.

I have not found any way to do this migration within the PyQt5 version
of the application because even trying to import the PyQt4 QString
fails on a conflict with PyQt5. The migrate-settings script can be
run once and will migrate the settings file to use a plain str instead
of QString so it should be able to load in PyQt5. No data will be
lost in this process.